Feat: Can now fetch and edit a note
