+++ src/components/notes/viewer.tsx
@@ -4,12 +4,13 @@ import remarkBreaks from 'remark-breaks'
 
 import type { Note } from '@/services/notes'
 
-const NoteViewer: React.FC<{ note: Note}> = props => {
+const NoteViewer: React.FC<{ note: Note }> = props => {
     return (
         <Markdown
             remarkPlugins={[remarkGfm, remarkBreaks]}
             className={`flex w-9/12 flex-col gap-2`}
             skipHtml={true}
+            // Parse \n as <br>
             components={{
                 h1: ({ node, ...props }) => <h1 className="text-2xl font-bold" {...props} />,
                 h2: ({ node, ...props }) => <h2 className="text-xl font-bold" {...props} />,
@@ -42,7 +43,12 @@ const NoteViewer: React.FC<{ note: Note}> = props => {
                 br: ({ node, ...props }) => <br {...props} />
             }}
         >
-            {props.note.content}
+            {/* Parse \n as <br> */}
+            {props.note.content.replaceAll(
+                '\\n',
+                `
+ `
+            )}
         </Markdown>
     )
 }
